Fix bad renaming
