+++ sites_faciles/blog/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class BlogConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name ="sites_faciles.blog"
+    label = "sites_faciles_blog"
